76/77 backend tests passing & 19/27 frontend tests passing

+++ back-end/src/db/migrations/20210123122225_createReservationsTable.js
@@ -12,6 +12,10 @@ exports.up = function (knex) {
   });
 };
 
-exports.down = function (knex) {
-  return knex.schema.dropTable("reservations");
+exports.down = function(knex) {
+  return knex.schema.hasTable('reservations').then(function(exists) {
+      if (exists) {
+          return knex.schema.dropTable('reservations');
+      }
+  });
 };
